# openapi3.yaml
openapi: 3.0.1
info:
  title: 3d-middleware
  description: Validating and routing requests
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /ingestion:
    post:
      operationId: createModel
      summary: Invokes a new model ingestion flow
      tags:
        - ingestion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/payload'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/job'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '5XX':
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /update:
    post:
      operationId: updateMetadata
      summary: Updates metadata of a model
      tags:
        - update
      requestBody:
        content:
          application/json:
            schema:
              $ref: './Schema/3dCatalog/3dMetadata.yaml#/components/schemas/updatePayload'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: './Schema/3dCatalog/3dMetadata.yaml#/components/schemas/metadata'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '5XX':
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    payload:
      type: object
      description: 3d model payload
      required:
        - modelPath
        - tilesetFilename
        - metadata
      properties:
        modelPath:
          type: string
          description: Model files location path
          example: '\\domtest\mimi\archi\maz\silver\libot\NewYork'
        tilesetFilename:
          type: string
          description: Model tileset filename
          example: 'tileset.json'
        metadata:
          $ref: './Schema/3dCatalog/3dMetadata.yaml#/components/schemas/basic-payload'
      additionalProperties: false
    job:
      type: object
      description: job for ingestion 3d model
      required:
        - jobId
        - status
      properties:
        jobId:
          type: string
          format: uuid
          description: Ingestion job identifier
        status:
          type: string
          enum:
            - Completed
            - In-Progress
            - Pending
            - Failed
            - Expired
            - Aborted